Codeforces Round #630 (Div. 2)
A) a bit tricky when b - a == 0. You need to move 1 square in at least one direction if (b!=0 || a !=0). Movement is necessary in left if b - a < 0 or in right, if b - a > 0

B) It is not possible to construct a composite number which is made of 2 different prime numbers greater than the 11th prime(At least)
   Use those 11 primes to categorise each composite number into a color

C) find frequency of all alphabets for all the (n / k) segments. Find the minimum number of operations to make ith and (n - i - i)th alphabet the same char using brute force.

D) A 3x3 matrix is suficient to execute the task. We'll aim for zero score obtained using algorithm given, and 'k' score using the best approach.
   Find entries of a matrix such that using the algo, score becomes zero and using the best approach, score becomes k
   Example k = 2(10)
   m[][] = 111 010 010
	   111 100 111
	   100 100 010